refactor: rename _enrichment.py to _enricher.py
